python/requirements.txt
setuptools==50.3.2
grpcio==1.43.0
grpcio-tools==1.43.0
grpcio-health-checking==1.43.0
google-api-core==1.27.0
protobuf==3.18.1
gapic-google-cloud-speech-v1==0.15.3
grpc-google-cloud-speech-v1==0.8.1
proto-google-cloud-speech-v1==0.15.3
google-auth==1.21.1
google-auth-httplib2==0.0.3
google-cloud-core==1.0.2
google-cloud-speech==1.0.0
googleapis-common-protos==1.6.0
httplib2==0.14.0
oauth2client==2.0.0
pydub==0.23.1
pyaudio==0.2.11
